BF: restructure of the player-code into separate folder
